add rights check and auth requests

+++ backend/bibliophilia/books/api.py
@@ -1,7 +1,9 @@
+import json
 import logging
 
-from fastapi import APIRouter, Query
+from fastapi import APIRouter, Query, HTTPException, status
 from fastapi import Response
+from starlette.requests import Request
 
 import backend.bibliophilia.books.settings as settings
 from backend.bibliophilia.books.domain.entity.facet import Facet
@@ -12,46 +14,66 @@ from backend.bibliophilia.books.domain.models.output import BookInfo, BookCard
 import backend.bibliophilia.books.dependencies as dependencies
 from typing import Optional, Set, AnyStr
 from fastapi import UploadFile
-from starlette.responses import FileResponse
+from starlette.responses import FileResponse, RedirectResponse
 
 from backend.bibliophilia.books.domain.models.input import Rights
+from backend.bibliophilia.books.domain.utils.security import check_book_right, check_is_publisher
 
 router = APIRouter()
 
 
 @router.post("/upload", response_model=Optional[int])
-def handle_create_book(book: BookCreateInfo,
+def handle_create_book(request: Request,
+                       book: BookCreateInfo,
                        response: Response):
+    if request.session.get('user') is None:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Please login to upload books")
     book, response.status_code = dependencies.book_service.create_book(BookCreate(**book.dict()))
     logging.info(f"Book created: {book.idx}")
     return book.idx
 
 
 @router.post("/image/upload")
-def handle_upload_image(book_idx: str,
+def handle_upload_image(request: Request,
+                        book_idx: str,
                         image: Optional[UploadFile],
                         response: Response):
+    if request.session.get('user') is None:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Please login to upload image")
+    if not check_is_publisher(int(book_idx), request.session.get('user').get('email')):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Only the publisher can upload an image")
     _, response.status_code = dependencies.book_service.create_image(ImageFileSave(book_idx=book_idx,
                                                                                    image=image))
 
 
 @router.post("/file/upload")
-def handle_upload_file(book_idx: str,
-                        file: Optional[UploadFile],
-                        response: Response):
+def handle_upload_file(request: Request,
+                       book_idx: str,
+                       file: Optional[UploadFile],
+                       response: Response):
+    if request.session.get('user') is None:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Please login to upload file")
+    if not check_is_publisher(int(book_idx), request.session.get('user').get('email')):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Only the publisher can upload a file")
     response.status_code = dependencies.book_service.create_file(BookFileSave(book_idx=book_idx,
                                                                               file=file))
 
 
 @router.post("/add-rights")
-def handle_add_rights(book_idx: int, user_idx: int, rights: Rights):
-    ### проверка, что user - publisher через request
+def handle_add_rights(request: Request, book_idx: int, user_idx: int, rights: Rights):
+    if request.session.get('user') is None:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Please login to add rights")
+    if not check_is_publisher(book_idx, request.session.get('user').get('email')):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Only the publisher can add rights")
     dependencies.book_service.add_rights(book_idx=book_idx, rights=rights, user_idx=user_idx)
 
 
 @router.delete("/delete-rights")
-def handle_delete_rights(book_idx: int):
-    ### проверка, что user - publisher через request
+def handle_delete_rights(request: Request, book_idx: int):
+    if request.session.get('user') is None:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Please login to delete rights")
+    if not check_is_publisher(book_idx, request.session.get('user').get('email')):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Only the publisher can add rights")
     dependencies.book_service.delete_rights(book_idx)
 
 
@@ -64,11 +86,18 @@ def handle_get_book_info(idx: int):
 
 
 @router.get("/search/", response_model=list[BookCard])
-def handle_search_books(q: str = Query("", title="Query string"),
+def handle_search_books(request: Request,
+                        q: str = Query("", title="Query string"),
                         page: int = Query(1, title="Page number")):
     books = dependencies.search_service.search(query=q, page=page)
+    show_books = []
+    for book in books:
+        access = check_book_right(book.idx, request.session.get('user'), "search")
+        if access == True:
+            show_books.append(book)
+
     logging.info(f"Books Founded: {len(books)}\n{books}")
-    return books
+    return show_books
 
 
 @router.get("/search/facets", response_model=Set[Facet])
@@ -85,7 +114,10 @@ def handle_get_hints(q: str = Query("", title="Query"), facet: Facet = Query(Non
 
 
 @router.get("/download/")
-def handle_download_bookfile(idx: int, book_format: str):
+def handle_download_bookfile(request: Request, idx: int, book_format: str):
+    access = check_book_right(idx, request.session.get('user'), "download")
+    if not access:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You don't have rights for download this book!")
     book = dependencies.book_service.read_book(idx=idx)
     bookfile = dependencies.book_service.read_bookfile(idx=idx, file_format=FileFormat.get_by_name(book_format))
     if bookfile and book:
